+++ src/main/java/com/roadtax/repositories/TaxPaymentRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.rest.core.annotation.RepositoryRestResource;
 
 @RepositoryRestResource(collectionResourceRel = "taxPayment", path="taxPayment")
-public interface TaxPaymentRepository extends JpaRepository<TaxPayment, Long>
+public interface TaxPaymentRepository extends JpaRepository<TaxPayment, Integer>
 {
 }
